Add mysql_host and qpid_host parameters
This patch allows the addresses for both mysql and qpid servers to set
independently of the controller private IP address. This provides
greater flexibility. For example, if the controller, mysql and qpid
servers had different virtual IP addresses. Note that the default
value will still be the controller private IP address, so unless these
new parameters are overridden there is no change in behavior.