Fix errors on inactive nodes.
Previously the inactive nodes in a cluster would still attempt to call the
database class and create their DBs and users, which of course, would fail,
since there was no mysql running on those nodes.  This patch simply leverages
the same info as our (soon to be cleaned up) execs to determine if those
database calls should happen or not.

However, it currently has the problem that using the fact fails on the first
run, since mysql is not yet running.  This still needs more work, but at least
gets us closer to working, as nodes should eventually show success instead of
always failing if they are not the active mysql host.