Remove password reset from foreman_server.sh
The line is not necessary to create an admin account in a new
installation, admin account should be created anyway. However it could
reset password back to the default one in existing installations,
which would be undesirable.

Info about the default password and how to change it has been added to
the instructions at the end of the script.