Fix neutron keystone auth config
Neutron::keystone::auth was resourced twice when $neutron == 'true',
and this caused puppet error. I first tried to remove it from
quickstack::controller and keep it in quickstack::neutron::controller,
but the openstack::keystone class seems to require
neutron_user_password even when neutron is disabled, so we'll better
configure it from quickstack::controller and just use the `neutron`
boolean parameter to enable or disable creating keystone entries for
neutron.

Also, wrap $neutron variable reading into str2bool as it is in other
places.