Fix enable/disable of heat_cfg and heat_cloudwatch
BZ#1039661 - This patch modifies the heat_controller.pp manifest such
that the heat_cfn and heat_cloudwatch services are enabled/disabled
per the heat_cfn and heat_cloudwatch boolean parameters. Previously,
once either boolean had been set to true the services would be running
on the controller, but reverting either parameter to false would not
stop/disable the service.

This patch also fixes the call to the heat::keystone::auth class such
that the heat_cfn endpoints are always created with the correct IP
addresses.

https://bugzilla.redhat.com/show_bug.cgi?id=1039661