BZ #1042933 - neutron-server fails to start
https://bugzilla.redhat.com/show_bug.cgi?id=1042933

Neutron-server fails to start when deploying Neutron controller host group via
Foreman.

This is due to a new neutron-db-manage command that needs to be run, but is not
yet included in puppet-neutron.  This patch mimics what was done in
packstack[1], fixing this until the proper fix is added to the puppet module we
depend on.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1037675