BZ #1017281 - Add support for ML2 Core Plugin.
https://bugzilla.redhat.com/show_bug.cgi?id=1017281

The ml2 code in puppet-neutron seems a bit buggy, so this currently includes
some workarounds, like installing the ml2 package [1], which puppet-neutron
should do.  There also is some extra work required to do the configuration for
the mechanism_drivers, which seem to require installation of their related
packages as well, to get the right configuration files.

[1] Fixed upstream in https://review.openstack.org/#/c/66413/