Configure qpid_hostname for Controller host groups
The default controller common manifest assumes that qpid is
running locally on the hypervisor. This may or may not be the case
and therefore it should take the '$qpid_host' parameter used within
other manifests. Without this change, Nova attempts to use
localhost as it's qpid broker, which obviously breaks if using
multiple controller hosts.