Fencing options added to quickstack::pacemaker::common
Set fencing_type to 'disabled' to set the stonith-enabled pacemaker
propety to false.

Set fencing_type to 'fence_xvm' to add fence_xvm resources.  If
fence_xvm is true, also set:

  fence_xvm_clu_iface -- the interface the cluster IP resides on
                         (as known by "pcs status")
  fence_xvm_manage_key_file -- true or false, whether or not to
                               write /etc/cluster/fence_xvm.key
  fence_xvm_key_file_password -- if the above is true, the contents
                                 of /etc/cluster/fence_xvm.key

  Note that `hostname -s` must match the name the name of of the guest
  in virsh for fence_xvm to work here.

Set fencing_type to 'ipmilan' to add a fence_ipmilan stonith resource.
In this case, the other relevant parameters are:

fence_ipmilan_address -- the IP of the ipmilan fence device
fence_ipmilan_username -- the username for the fence device, if needed
fence_ipmilan_password -- the password for the fence device, if needed
fence_ipmilan_interval -- interval passed to the pcs stonith create
                          command, e.g. '30s'