BZ #1073087 - [RFE] Use subscription-manager
https://bugzilla.redhat.com/show_bug.cgi?id=1073087

Use subscription-manager by default instead of rhnreg_ks to subscribe
the system for Red Hat repositories.

Only parameters for username/password registration are present by
default, but more subscription options are available (including the
original rhnreg_ks). Those parameters are commented out, not populated
by default, not to confuse users what params to fill in. (Creating a
parameter with an empty string as value is not allowed.)

This commit depends on Foreman having a redhat_register snippet that
supports subscription-manager and also has a workaround for correctly
enabling repositories on RHEL 6.4 [1].

[1] https://github.com/theforeman/community-templates/commit/6a66f3908b2c8729d311507ec5e56ee2f34e01c7