BZ #1086344 - Add Cinder service to HA hostgroup
https://bugzilla.redhat.com/show_bug.cgi?id=1086344

Adds Cinder to the HA controller host group. So far only iscsi backend
is supported. Note that iscsi backend requires cinder-volumes LVM
volume group to exist on the target machine. I don't think this will
be a handy configuration for HA host group in production (data saved
in the LVM iscsi backend is not really HA), but it is handy for
testing. The installation of cinder-volume service on the HA
controller is optional, switched off by default.

Gluster support should follow and it should also be made the default
backend.

(Also enables creation of Cinder DB in the hamysql::singlenodetest
class.)