BZ #1062699 - Cinder + GlusterFS mount options on HA controller
https://bugzilla.redhat.com/show_bug.cgi?id=1062699

Add GlusterFS support to Cinder on HA Controller and make it possible
to enter also mount options for it. The mount options can be entered
by setting glusterfs_shares parameter to e.g.:

    ["gluster1:/cinder -o backupvolfile-server=gluster2"]

Changed cinder-volume related parameters and default values on the HA
Controller - especially important is defaulting glusterfs_shares to []
so that Foreman registers the parameter type as array.