Add *_network params and logic.
The purpose of this is to allow a user to specify what they know, which
sometimes is not the nic (which may not be known to them at setup time until
after the machine is provisioned).  They should be able to provide the network
they know a nic should be in though, and we can use this to figure out which nic
got attached to that network.

(cherry picked from commit 4e4a138288fdaf6336758f3944da4a731571919b)