ha controller: add fence_ipmilan_host_to_address param to commmon
This is a list of pairs that determines what address to use
when creating a stonith ipmilan device in pacemaker.

E.g., given a fence_ipmilan_host_to_address that looks like:
["192.168.200.10:10.10.10.10", "192.168.200.20:10.10.10.20",
"192.168.200.30:10.10.10.30"],

the cluster node 192.168.200.10 should use 10.10.10.10 as its ipmilan
fence address.

Note that the "address" param must be blank when providing the
"fence_ipmilan_host_to_address" param.