ensure pcs resource create commands run on just one node
We want to avoid getting unlucky and running the exact "pcs resource
create" command on two nodes at once which can destabilize pacemaker
(although this is not a problem upstream).  So, this commit adds a
parameter, cluster_control_ip as the one IP to run "pcs resource
create" commands from.