BZ #1103315 - Openstack firewall rules are not enabled after reboot.
https://bugzilla.redhat.com/show_bug.cgi?id=1103315

There was an additional issue where iptables rules were attempted to be
persisted before the service had been started.  We need to make sure we call the
base firewall class on our side, which is what this patch does, and then once the
firewall module gets the fix needed on that side[1], everything should work more
cleanly.

[1] https://github.com/puppetlabs/puppetlabs-firewall/pull/367