Change network_floating_range default
The defaults for the network_fixed_range and network_floating_range
parameters were the same.  Even if this is just defaults that we
expect to be overridden, they should not be the same as that would be
a broken configuration.  They must be separate networks.  Go ahead and
change the defaults of one of them (floating in this case) to be
something that would work.