Fix File dep issue for gluster on compute.
gluster::mount::base pulls in gluster::vardir for temorary write of files, but
that in turn needs _some_ tmp dir set up.  The gluster module is set up so the
admin can set up whatever they prefer for that tmp dir, but the one in
puppet::vardir is what we will use for now, so this simply makes sure that dir
is there.