BZ #1123318 - HA: set max_retries=-1 for accessing the DB
https://bugzilla.redhat.com/show_bug.cgi?id=1123318

Services will now continue to retry accessing the db on startup if it
is not available, as the max_retries configuration value in the
openstack .conf files is now set to -1.

From BZ:
"HA: All services are missing max_retries=-1 settings for accessing the DB.

This means that if the DB is not started before the service, the
service will fail to start. This behaviour is currently mitigated by:
start-failure-is-fatal: false, but it is technically incorrect."