BZ#113272 - Cinder volume active/active should be avoided
https://bugzilla.redhat.com/show_bug.cgi?id=1132722

There's a bug reported against Cinder that if multiple cinder-volume
services are configured with the same backend storage and one
cinder-volume service goes down, its volumes are stranded - the other
cinder-volume services do not take over the management of these
volumes. Details on Launchpad [1].

The workaround is to set the "host" configuration option to the same
value for all cinder-volume services, so that they all appear as one,
and use active/passive HA for cinder-volume.

The side effect is that if we use iSCSI backend with local LVM data
storage for Cinder on the HA controllers, we can only use the disk
space on one of the controllers, not all (cinder-volume service will
run on only one controller). This should not be a problem because it
is not an expected production setup with HA controller - Cinder
backend setup using local LVM data storage on controllers wouldn't
provide HA anyway. If the user wants to scale out LVM data storage
backend, they can still use the Block Storage host group. This commit
only affects HA controller.

[1] https://bugs.launchpad.net/cinder/+bug/1322190