BZ #1124484 - Puppet fails on networker after br-ex is configured.
https://bugzilla.redhat.com/show_bug.cgi?id=1124484

Previously, once the IP was moved to the bridge, subsequent puppet runs would
fail because puppet would (rightly) find no ip on the ovs_tunnel_iface.  This
patch updates the find_ip finction to optionally take a list of interfaces to
check.  In the case of the networker node, we use the external_network_bridge to
check for IP as well, if set.