switch arrow type between execs
The "~>" arrow allows the thing on the right to be executed even if
the thing on the left is skipped.  This was bad for the two dep's
updated in this PR; e.g. we would see a glance db_sync failure even
if we saw something like:
  Exec[i-am-glance-vip-OR-glance-is-up-on-vip]: Skipping because of failed dependencies
when instead the db_sync shouldn't have even been attempted.