Removed standard function each
Function 'each' requires parser=future which we don't want to use

This patch replaces the class include with a custom function binding the
classes. This is the same approach as hiera_include which we used before
adding the Scenes layer.

Also changed quickstack::scene for more clarity in the output.