Fix setting of auth_url from glance cache.
As noted in https://bugzilla.redhat.com/show_bug.cgi?id=1174389#c10 ,
glance-cache.conf:auth_url was set to http://localhost:5000/v2.0.  This is
because the glance::api module does not build this url from the auth_host we
were already passing in.  So, at least for now, we do this on the quickstack
side.