BZ #1168755 - rabbitmqctl failure while setting HA policy
https://bugzilla.redhat.com/show_bug.cgi?id=1168755

The policies are set in the clustered database and replicated across
the cluster.  As such, it really only needs to be set on the first
node, and then it will replicate to the others when they join the
cluster.  I think something like the following happens:

- First node starts, sets HA policy

- Second node "starts".  My suspicion is that there is probably some
  cluster state sync happening in the background after the systemd
  unit succeeds.

- Second node attempts to set HA policy (unnecessarily) but cannot
  connect to local rabbit node due to race condition with clustering
  and state synchronization.

- Explosions and sadness, puppet errors.

This change declares the exec only for the first node, so additional
cluster members will just inherit the policy from the cluster state.